driver: display: Add support for bbc:microbit v2
Add support display driver for new bbc:microbit v2 board.

This PR decouples the dependence of logic and hardware,
and has achieved maximum code reuse.
