+++ drivers/display/Kconfig.microbit
@@ -5,7 +5,7 @@
 
 config MICROBIT_DISPLAY
 	bool "BBC micro:bit 5x5 LED Display support"
-	depends on BOARD_BBC_MICROBIT
+	depends on BOARD_BBC_MICROBIT || BOARD_BBC_MICROBIT_V2
 	depends on PRINTK
 	depends on GPIO
 	help
